[feat] #57 mapping 함수 구현

+++ presentation/src/main/java/com/dpm/sixpack/presentation/common/util/UiStateMapperUtil.kt
@@ -34,14 +34,14 @@ fun calculatePace(
     return String.format("%02d'%02d", minutes, seconds)
 }
 
-fun formatPaceToString(paceInSeconds: Int): String {
+fun formatPaceToString(paceInSeconds: Long): String {
     val minutes = paceInSeconds / 60
     val seconds = paceInSeconds % 60
 
     return String.format("%02d'%02d", minutes, seconds)
 }
 
-fun formatDistanceToKm(distanceInMeters: Int): String {
+fun formatDistanceToKm(distanceInMeters: Long): String {
     if (distanceInMeters < 1000) return "${distanceInMeters}m"
     val km = distanceInMeters / 1000.0
     val formattedString = String.format("%.2f", km)
